Fix round-robin: integer idle-time bucketing caused first-registered node to always win
score_node bucketed idle time with int(idle_seconds / 10), which rounded
nodes that finished tasks within the same 10-second window to the same
score. Python dict preserves insertion order so the first-registered node
always won the tie, monopolising primary slots indefinitely.

Switching to float division means millisecond differences in last_task_time
break ties correctly. Confirmed in smoke test: 4 tasks split 2/2 between
two nodes with earnings of 20.53 each.

+++ protocol/orchestrator.py
@@ -161,9 +161,12 @@ def score_node(node: NodeConnection, model: str) -> int:
     # Hardware (gpu / vram_gb) intentionally not scored — self-reported and
     # unverified. See module docstring and docs/VERIFICATION.md build-now #3.
 
-    # Round-robin tiebreaker: prefer the node that has been idle longest
+    # Round-robin tiebreaker: prefer the node that has been idle longest.
+    # Use a float ratio so even millisecond differences break ties — prevents
+    # the same node from monopolising primary slots when all nodes finish
+    # at nearly the same time.
     idle_seconds = time.time() - node.last_task_time
-    score += min(int(idle_seconds / 10), 5)
+    score += min(idle_seconds / 10.0, 5.0)
 
     return score
